# HPACK Encoder headerFields improvements