@normanmaurer 2016-06-29
LGTM

> Am 29.06.2016 um 03:16 schrieb Scott Mitchell <notifications@github.com>:
> 
> Motivation:
> HPACK Encoder has a data structure which is similar to a previous version of DefaultHeaders. Some of the same improvements can be made.
> 
> Motivation:
> 
> Enforce the restriction that the Encoder's headerFields length must be a power of two so we can use masking instead of modulo
> Use AsciiString.hashCode which already has optimizations instead of having yet another hash code algorithm in Encoder
> Result:
> Fixes #5357
> 
> You can view, comment on, or merge this pull request online at:
> 
>   https://github.com/netty/netty/pull/5466
> 
> Commit Summary
> 
> HPACK Encoder headerFields improvements
> File Changes
> 
> M codec-http2/src/main/java/io/netty/handler/codec/http2/internal/hpack/Encoder.java (65)
> M codec-http2/src/test/java/io/netty/handler/codec/http2/internal/hpack/TestCase.java (2)
> M codec/src/main/java/io/netty/handler/codec/DefaultHeaders.java (10)
> M common/src/main/java/io/netty/util/AsciiString.java (2)
> M microbench/src/main/java/io/netty/microbench/http2/internal/hpack/EncoderBenchmark.java (8)
> Patch Links:
> 
> https://github.com/netty/netty/pull/5466.patch
> https://github.com/netty/netty/pull/5466.diff
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub, or mute the thread.
>