@Scottmitch 2016-06-30
```
=== with PR

Benchmark                (duplicates)  (limitToAscii)  (maxTableSize)  (sensitive)  (size)   Mode  Cnt        Score        Error  Units
EncoderBenchmark.encode          true            true            4096         true   SMALL  thrpt    5   356927.366 ±  56001.273  ops/s
EncoderBenchmark.encode          true            true            4096         true  MEDIUM  thrpt    5    47351.315 ±   8528.331  ops/s
EncoderBenchmark.encode          true            true            4096         true   LARGE  thrpt    5     2720.457 ±    193.968  ops/s
EncoderBenchmark.encode          true            true            4096        false   SMALL  thrpt    5  1437241.861 ± 132873.605  ops/s
EncoderBenchmark.encode          true            true            4096        false  MEDIUM  thrpt    5   849782.491 ±  86691.285  ops/s
EncoderBenchmark.encode          true            true            4096        false   LARGE  thrpt    5   144427.332 ±   7858.671  ops/s
EncoderBenchmark.encode          true           false            4096         true   SMALL  thrpt    5  2830273.724 ± 140489.445  ops/s
EncoderBenchmark.encode          true           false            4096         true  MEDIUM  thrpt    5   488952.888 ±  43357.410  ops/s
EncoderBenchmark.encode          true           false            4096         true   LARGE  thrpt    5    36376.959 ±   3110.164  ops/s
EncoderBenchmark.encode          true           false            4096        false   SMALL  thrpt    5  4482576.931 ± 137792.996  ops/s
EncoderBenchmark.encode          true           false            4096        false  MEDIUM  thrpt    5  1435410.956 ±  84211.360  ops/s
EncoderBenchmark.encode          true           false            4096        false   LARGE  thrpt    5   194170.322 ±  19504.159  ops/s
EncoderBenchmark.encode         false            true            4096         true   SMALL  thrpt    5   368044.726 ±  34911.272  ops/s
EncoderBenchmark.encode         false            true            4096         true  MEDIUM  thrpt    5    44510.592 ±   1875.679  ops/s
EncoderBenchmark.encode         false            true            4096         true   LARGE  thrpt    5     2489.113 ±    127.246  ops/s
EncoderBenchmark.encode         false            true            4096        false   SMALL  thrpt    5   338468.082 ±  12103.634  ops/s
EncoderBenchmark.encode         false            true            4096        false  MEDIUM  thrpt    5    40525.168 ±   2354.368  ops/s
EncoderBenchmark.encode         false            true            4096        false   LARGE  thrpt    5     2365.145 ±    160.677  ops/s
EncoderBenchmark.encode         false           false            4096         true   SMALL  thrpt    5  2663665.857 ± 364377.319  ops/s
EncoderBenchmark.encode         false           false            4096         true  MEDIUM  thrpt    5   478849.332 ±  41604.573  ops/s
EncoderBenchmark.encode         false           false            4096         true   LARGE  thrpt    5    35103.798 ±   1318.136  ops/s
EncoderBenchmark.encode         false           false            4096        false   SMALL  thrpt    5  2379261.970 ±  57179.624  ops/s
EncoderBenchmark.encode         false           false            4096        false  MEDIUM  thrpt    5   411287.298 ±  22104.069  ops/s
EncoderBenchmark.encode         false           false            4096        false   LARGE  thrpt    5    30752.950 ±   2407.385  ops/s

=== before PR

Benchmark                (duplicates)  (limitToAscii)  (maxTableSize)  (sensitive)  (size)   Mode  Cnt        Score        Error  Units
EncoderBenchmark.encode          true            true            4096         true   SMALL  thrpt    5   352301.863 ±  54569.696  ops/s
EncoderBenchmark.encode          true            true            4096         true  MEDIUM  thrpt    5    47410.641 ±   4530.055  ops/s
EncoderBenchmark.encode          true            true            4096         true   LARGE  thrpt    5     2841.511 ±    113.221  ops/s
EncoderBenchmark.encode          true            true            4096        false   SMALL  thrpt    5  1277058.234 ±  74586.996  ops/s
EncoderBenchmark.encode          true            true            4096        false  MEDIUM  thrpt    5   414200.033 ±  30139.221  ops/s
EncoderBenchmark.encode          true            true            4096        false   LARGE  thrpt    5    66775.632 ±   3148.258  ops/s
EncoderBenchmark.encode          true           false            4096         true   SMALL  thrpt    5  2747344.162 ± 193883.568  ops/s
EncoderBenchmark.encode          true           false            4096         true  MEDIUM  thrpt    5   487397.227 ±  42936.079  ops/s
EncoderBenchmark.encode          true           false            4096         true   LARGE  thrpt    5    35687.226 ±   3862.617  ops/s
EncoderBenchmark.encode          true           false            4096        false   SMALL  thrpt    5  3124704.076 ±  90154.430  ops/s
EncoderBenchmark.encode          true           false            4096        false  MEDIUM  thrpt    5   720691.100 ±  31043.752  ops/s
EncoderBenchmark.encode          true           false            4096        false   LARGE  thrpt    5    74676.306 ±   3189.223  ops/s
EncoderBenchmark.encode         false            true            4096         true   SMALL  thrpt    5   366475.332 ±  38861.898  ops/s
EncoderBenchmark.encode         false            true            4096         true  MEDIUM  thrpt    5    44272.063 ±   2756.406  ops/s
EncoderBenchmark.encode         false            true            4096         true   LARGE  thrpt    5     2504.013 ±    536.302  ops/s
EncoderBenchmark.encode         false            true            4096        false   SMALL  thrpt    5   527139.189 ±  49906.776  ops/s
EncoderBenchmark.encode         false            true            4096        false  MEDIUM  thrpt    5    39418.695 ±   2531.731  ops/s
EncoderBenchmark.encode         false            true            4096        false   LARGE  thrpt    5     2259.957 ±     73.728  ops/s
EncoderBenchmark.encode         false           false            4096         true   SMALL  thrpt    5  2623078.195 ±  56393.871  ops/s
EncoderBenchmark.encode         false           false            4096         true  MEDIUM  thrpt    5   446502.577 ±  32295.207  ops/s
EncoderBenchmark.encode         false           false            4096         true   LARGE  thrpt    5    34913.448 ±   2797.255  ops/s
EncoderBenchmark.encode         false           false            4096        false   SMALL  thrpt    5  1417427.482 ±  45159.113  ops/s
EncoderBenchmark.encode         false           false            4096        false  MEDIUM  thrpt    5   200201.677 ±   9379.406  ops/s
EncoderBenchmark.encode         false           false            4096        false   LARGE  thrpt    5    17524.141 ±   1062.549  ops/s
```